+++ src/msb_arch/__init__.py
@@ -4,6 +4,7 @@ Mega-Super-Base (MSB) architecture package.
 """
 
 from .base import Serializable, BaseEntity, BaseContainer
+from .base.serializable import cache_statistics
 from .errors import (AttributeNotFoundError,
                      ConstraintError,
                      DispatchError,
@@ -20,7 +21,10 @@ from .errors import (AttributeNotFoundError,
                      TypeValidationError,
                      UnknownAttributeError,
                      ValidationError)
+from .interceptors import RequestJournal, RequestMetrics
+from .protocols import Interceptor, MethodProvider
 from .super import Super, Project
+from .super.builtins import Configurator, Inspector
 from .mega import Manipulator
 from .results import MethodResults
 from .utils import logger, setup_logging
@@ -36,6 +40,8 @@ __all__ = ["Serializable", "BaseEntity", "BaseContainer", "Super", "Project", "M
            "MSBError", "ValidationError", "TypeValidationError", "ConstraintError", "UnknownAttributeError", "ItemNameError",
            "DuplicateNameError", "ResolutionError", "NotFoundError", "AttributeNotFoundError", "SerializationError",
            "OperationError", "RegistrationError", "DispatchError", "RequestError", "HandlerError",
-           "Constraint", "Positive", "NonNegative", "NonZero", "NonEmpty", "Range", "Predicate"]
+           "Constraint", "Positive", "NonNegative", "NonZero", "NonEmpty", "Range", "Predicate",
+           "MethodProvider", "Interceptor", "Inspector", "Configurator",
+           "RequestMetrics", "RequestJournal", "cache_statistics"]
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"